implemented strategy for the shapes as well.

+++ Calculator/Interfaces/IShape.cs
@@ -2,9 +2,9 @@
 {
     public interface IShape
     {
-        public decimal Width { get; set; }
-        public decimal Height { get; set; }
-        public decimal CalculateArea(decimal width, decimal height);
-        public decimal CalculateCircumference(decimal width, decimal height);
+        public double Width { get; set; }
+        public double Height { get; set; }
+        public double CalculateArea(double width, double height);
+        public double CalculateCircumference(double width, double height);
     }
 }
